refactor: split builtin signatures by domain
