+++ libs/mql-interpreter/src/libs/signatures/indicators.ts
@@ -1,0 +1,441 @@
+import { BuiltinSignaturesMap } from "./types";
+
+// https://docs.mql4.com/indicators
+export const indicatorsBuiltinSignatures: BuiltinSignaturesMap = {
+  iAC: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Accelerator Oscillator",
+  },
+  iAD: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Accumulation/Distribution",
+  },
+  iADX: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "period", type: "int" },
+      { name: "applied_price", type: "int" },
+      { name: "mode", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Average Directional Index",
+  },
+  iAlligator: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "jaw_period", type: "int" },
+      { name: "jaw_shift", type: "int" },
+      { name: "teeth_period", type: "int" },
+      { name: "teeth_shift", type: "int" },
+      { name: "lips_period", type: "int" },
+      { name: "lips_shift", type: "int" },
+      { name: "ma_method", type: "int" },
+      { name: "applied_price", type: "int" },
+      { name: "mode", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Alligator",
+  },
+  iAO: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Awesome Oscillator",
+  },
+  iATR: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "period", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Average True Range",
+  },
+  iBands: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "period", type: "int" },
+      { name: "deviation", type: "double" },
+      { name: "bands_shift", type: "int" },
+      { name: "applied_price", type: "int" },
+      { name: "mode", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Bollinger Bands®",
+  },
+  iBandsOnArray: {
+    args: [
+      { name: "array", type: "double[]" },
+      { name: "total", type: "int" },
+      { name: "period", type: "int" },
+      { name: "deviation", type: "double" },
+      { name: "bands_shift", type: "int" },
+      { name: "mode", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Calculation of Bollinger Bands® indicator on data, stored in a numeric array",
+  },
+  iBearsPower: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "period", type: "int" },
+      { name: "applied_price", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Bears Power",
+  },
+  iBullsPower: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "period", type: "int" },
+      { name: "applied_price", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Bulls Power",
+  },
+  iBWMFI: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Market Facilitation Index by Bill Williams",
+  },
+  iCCI: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "period", type: "int" },
+      { name: "applied_price", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Commodity Channel Index",
+  },
+  iCCIOnArray: {
+    args: [
+      { name: "array", type: "double[]" },
+      { name: "total", type: "int" },
+      { name: "period", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description:
+      "Calculation of Commodity Channel Index indicator on data, stored in a numeric array",
+  },
+  iCustom: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "name", type: "string" },
+      { name: "args", type: "any", variadic: true },
+    ],
+    returnType: "double",
+    description: "Custom indicator",
+  },
+  iDeMarker: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "period", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "DeMarker",
+  },
+  iEnvelopes: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "ma_period", type: "int" },
+      { name: "ma_shift", type: "int" },
+      { name: "ma_method", type: "int" },
+      { name: "applied_price", type: "int" },
+      { name: "deviation", type: "double" },
+      { name: "mode", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Envelopes",
+  },
+  iEnvelopesOnArray: {
+    args: [
+      { name: "array", type: "double[]" },
+      { name: "total", type: "int" },
+      { name: "ma_period", type: "int" },
+      { name: "ma_shift", type: "int" },
+      { name: "ma_method", type: "int" },
+      { name: "deviation", type: "double" },
+      { name: "mode", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Calculation of Envelopes indicator on data, stored in a numeric array",
+  },
+  iForce: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "period", type: "int" },
+      { name: "ma_method", type: "int" },
+      { name: "applied_price", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Force Index",
+  },
+  iFractals: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "mode", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Fractals",
+  },
+  iGator: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "jaw_period", type: "int" },
+      { name: "jaw_shift", type: "int" },
+      { name: "teeth_period", type: "int" },
+      { name: "teeth_shift", type: "int" },
+      { name: "lips_period", type: "int" },
+      { name: "lips_shift", type: "int" },
+      { name: "ma_method", type: "int" },
+      { name: "applied_price", type: "int" },
+      { name: "mode", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Gator Oscillator",
+  },
+  iIchimoku: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "tenkan_sen", type: "int" },
+      { name: "kijun_sen", type: "int" },
+      { name: "senkou_span_b", type: "int" },
+      { name: "mode", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Ichimoku Kinko Hyo",
+  },
+  iMA: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "period", type: "int" },
+      { name: "ma_shift", type: "int" },
+      { name: "ma_method", type: "int" },
+      { name: "applied_price", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Moving Average",
+  },
+  iMACD: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "fast_ema_period", type: "int" },
+      { name: "slow_ema_period", type: "int" },
+      { name: "signal_period", type: "int" },
+      { name: "applied_price", type: "int" },
+      { name: "mode", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Moving Averages Convergence-Divergence",
+  },
+  iMAOnArray: {
+    args: [
+      { name: "array", type: "double[]" },
+      { name: "total", type: "int" },
+      { name: "period", type: "int" },
+      { name: "ma_shift", type: "int" },
+      { name: "ma_method", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Calculation of Moving Average indicator on data, stored in a numeric array",
+  },
+  iMFI: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "period", type: "int" },
+      { name: "applied_price", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Money Flow Index",
+  },
+  iMomentum: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "period", type: "int" },
+      { name: "applied_price", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Momentum",
+  },
+  iMomentumOnArray: {
+    args: [
+      { name: "array", type: "double[]" },
+      { name: "total", type: "int" },
+      { name: "period", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Calculation of Momentum indicator on data, stored in a numeric array",
+  },
+  iOBV: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "applied_price", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "On Balance Volume",
+  },
+  iOsMA: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "fast_ema_period", type: "int" },
+      { name: "slow_ema_period", type: "int" },
+      { name: "signal_period", type: "int" },
+      { name: "applied_price", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Moving Average of Oscillator (MACD histogram)",
+  },
+  iRSI: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "period", type: "int" },
+      { name: "applied_price", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Relative Strength Index",
+  },
+  iRSIOnArray: {
+    args: [
+      { name: "array", type: "double[]" },
+      { name: "total", type: "int" },
+      { name: "period", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Calculation of Momentum indicator on data, stored in a numeric array",
+  },
+  iRVI: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "period", type: "int" },
+      { name: "mode", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Relative Vigor Index",
+  },
+  iSAR: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "step", type: "double" },
+      { name: "maximum", type: "double" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Parabolic Stop And Reverse System",
+  },
+  iStdDev: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "ma_period", type: "int" },
+      { name: "ma_shift", type: "int" },
+      { name: "ma_method", type: "int" },
+      { name: "applied_price", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Standard Deviation",
+  },
+  iStdDevOnArray: {
+    args: [
+      { name: "array", type: "double[]" },
+      { name: "total", type: "int" },
+      { name: "ma_period", type: "int" },
+      { name: "ma_shift", type: "int" },
+      { name: "ma_method", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Calculation of Standard Deviation indicator on data, stored in a numeric array",
+  },
+  iStochastic: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "k_period", type: "int" },
+      { name: "d_period", type: "int" },
+      { name: "slowing", type: "int" },
+      { name: "method", type: "int" },
+      { name: "price_field", type: "int" },
+      { name: "mode", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Stochastic Oscillator",
+  },
+  iWPR: {
+    args: [
+      { name: "symbol", type: "string" },
+      { name: "timeframe", type: "int" },
+      { name: "period", type: "int" },
+      { name: "shift", type: "int" },
+    ],
+    returnType: "double",
+    description: "Williams' Percent Range",
+  },
+};
